# capstone-implementation/capstone-implementation/openapi.yaml
openapi: 3.1.0
info:
  title: Nopbai Personal Loan Platform API
  description: Public contract for Nopbai Mobile Personal Loan. Realizes the three
    I-11 use cases (Submit and Decide Loan Application, Disburse Approved Loan Application,
    Recommend Limit Increase) and documents the in-scope Lab 3 contract rows C-01/C-02/C-03
    as simulated I-3 backing services.
  version: 1.0.0
paths:
  /loan-applications:
    post:
      tags:
      - Submit and Decide Loan Application
      summary: Submit And Decide
      operationId: submitAndDecideLoanApplication
      parameters:
      - name: x-customer-id
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Customer-Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubmitApplicationRequest'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubmitApplicationResponse'
        '422':
          description: CON.2 out-of-segment or CON.1 amount cap
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '503':
          description: CON.3 scoring timeout (controlled)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '401':
          description: CON.5 unauthenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
  /loan-applications/{application_id}/agreement:
    post:
      tags:
      - Disburse Approved Loan Application
      summary: Accept Agreement
      operationId: acceptLoanAgreement
      parameters:
      - name: application_id
        in: path
        required: true
        schema:
          type: string
          title: Application Id
      - name: x-customer-id
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Customer-Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplicationStateResponse'
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
          description: Unauthorized
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /loan-applications/{application_id}/decline:
    post:
      tags:
      - Submit and Decide Loan Application
      summary: Decline Offer
      operationId: declineLoanOffer
      parameters:
      - name: application_id
        in: path
        required: true
        schema:
          type: string
          title: Application Id
      - name: x-customer-id
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Customer-Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplicationStateResponse'
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
          description: Unauthorized
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /loan-applications/{application_id}/disbursement:
    post:
      tags:
      - Disburse Approved Loan Application
      summary: Disburse
      operationId: disburseApprovedLoanApplication
      parameters:
      - name: application_id
        in: path
        required: true
        schema:
          type: string
          title: Application Id
      - name: x-customer-id
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Customer-Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DisbursementRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DisbursementResponse'
        '422':
          description: CON.4 account validation failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '502':
          description: CON.4 posting/confirmation failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '401':
          description: CON.5 unauthenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
  /limit-increase-recommendations:
    post:
      tags:
      - Recommend Limit Increase
      summary: Recommend Limit Increase
      operationId: recommendLimitIncrease
      parameters:
      - name: x-customer-id
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Customer-Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LimitIncreaseRequest'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LimitIncreaseResponse'
        '422':
          description: CON.2 out-of-segment or CON.1 amount cap
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '401':
          description: CON.5 unauthenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
  /backing/credit-scoring-system/scores:
    post:
      tags:
      - 'C-01 backing: Credit Scoring System (mock)'
      summary: C01 Get Credit Score
      operationId: getCreditScore
      parameters:
      - name: x-caller
        in: header
        required: true
        schema:
          type: string
          description: Calling container (must be Credit Scoring Adapter)
          title: X-Caller
        description: Calling container (must be Credit Scoring Adapter)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreditScoreRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreditScoreResponse'
        '403':
          description: I-9 forbidden caller
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '504':
          description: CON.3 timeout/unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /backing/esb-integration-layer/disbursements:
    post:
      tags:
      - 'C-02 backing: ESB Integration Layer (mock)'
      summary: C02 Send Disbursement
      operationId: sendDisbursementAndAccounting
      parameters:
      - name: x-caller
        in: header
        required: true
        schema:
          type: string
          description: Calling container (must be Disbursement Adapter)
          title: X-Caller
        description: Calling container (must be Disbursement Adapter)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DisbursementAccountingRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DisbursementAccountingResponse'
        '403':
          description: I-9 forbidden caller
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /backing/core-banking/postings:
    post:
      tags:
      - 'C-03 backing: Core Banking (mock)'
      summary: C03 Post Disbursement
      operationId: postDisbursementAndAccounting
      parameters:
      - name: x-caller
        in: header
        required: true
        schema:
          type: string
          description: Calling container (must be ESB Integration Layer)
          title: X-Caller
        description: Calling container (must be ESB Integration Layer)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DisbursementAccountingRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DisbursementAccountingResponse'
        '403':
          description: I-9 forbidden caller (e.g. Mobile App direct)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
components:
  schemas:
    ApplicationStateResponse:
      properties:
        application_id:
          type: string
          title: Application Id
        state:
          type: string
          title: State
      type: object
      required:
      - application_id
      - state
      title: ApplicationStateResponse
    CreditScoreRequest:
      properties:
        customer_id:
          type: string
          title: Customer Id
      type: object
      required:
      - customer_id
      title: CreditScoreRequest
    CreditScoreResponse:
      properties:
        customer_id:
          type: string
          title: Customer Id
        credit_score:
          type: integer
          title: Credit Score
      type: object
      required:
      - customer_id
      - credit_score
      title: CreditScoreResponse
    DisbursementAccountingRequest:
      properties:
        application_id:
          type: string
          title: Application Id
        amount_vnd:
          type: integer
          title: Amount Vnd
        disbursement_reference:
          type: string
          title: Disbursement Reference
        simulate_posting_failure:
          type: boolean
          title: Simulate Posting Failure
          default: false
      type: object
      required:
      - application_id
      - amount_vnd
      - disbursement_reference
      title: DisbursementAccountingRequest
    DisbursementAccountingResponse:
      properties:
        reference:
          type: string
          title: Reference
        application_id:
          type: string
          title: Application Id
        amount_vnd:
          type: integer
          title: Amount Vnd
        confirmed:
          type: boolean
          title: Confirmed
      type: object
      required:
      - reference
      - application_id
      - amount_vnd
      - confirmed
      title: DisbursementAccountingResponse
    DisbursementRequest:
      properties:
        simulate_posting_failure:
          type: boolean
          title: Simulate Posting Failure
          default: false
      type: object
      title: DisbursementRequest
    DisbursementResponse:
      properties:
        application_id:
          type: string
          title: Application Id
        state:
          type: string
          title: State
        disbursement_reference:
          type: string
          title: Disbursement Reference
      type: object
      required:
      - application_id
      - state
      - disbursement_reference
      title: DisbursementResponse
    ErrorBody:
      properties:
        error_code:
          type: string
          title: Error Code
          examples:
          - CON.1
        state:
          anyOf:
          - type: string
          - type: 'null'
          title: State
          examples:
          - Rejected
        detail:
          type: string
          title: Detail
          examples:
          - amount cap / policy rejection (CON.1)
      type: object
      required:
      - error_code
      - detail
      title: ErrorBody
      description: Stable error body for every named alt / CON.*. `error_code` is
        the CON.* id.
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    LimitIncreaseRequest:
      properties:
        customer_id:
          type: string
          title: Customer Id
          examples:
          - CUST-001
        requested_amount_vnd:
          type: integer
          title: Requested Amount Vnd
          examples:
          - 80000000
        customer_age:
          type: integer
          title: Customer Age
          examples:
          - 30
        is_existing_salaried:
          type: boolean
          title: Is Existing Salaried
          examples:
          - true
      type: object
      required:
      - customer_id
      - requested_amount_vnd
      - customer_age
      - is_existing_salaried
      title: LimitIncreaseRequest
    LimitIncreaseResponse:
      properties:
        recommendation_id:
          type: string
          title: Recommendation Id
        loan_offer:
          $ref: '#/components/schemas/LoanOfferResponse'
      type: object
      required:
      - recommendation_id
      - loan_offer
      title: LimitIncreaseResponse
    LoanOfferResponse:
      properties:
        loan_offer_id:
          type: string
          title: Loan Offer Id
        application_id:
          type: string
          title: Application Id
        amount_vnd:
          type: integer
          title: Amount Vnd
        interest_rate:
          type: number
          title: Interest Rate
        kind:
          type: string
          title: Kind
      type: object
      required:
      - loan_offer_id
      - application_id
      - amount_vnd
      - interest_rate
      - kind
      title: LoanOfferResponse
    SubmitApplicationRequest:
      properties:
        customer_id:
          type: string
          title: Customer Id
          examples:
          - CUST-001
        requested_amount_vnd:
          type: integer
          title: Requested Amount Vnd
          examples:
          - 50000000
        customer_age:
          type: integer
          title: Customer Age
          examples:
          - 30
        is_existing_salaried:
          type: boolean
          title: Is Existing Salaried
          examples:
          - true
        payment_account_id:
          type: string
          title: Payment Account Id
          examples:
          - ACC-001
        simulate_scoring_timeout:
          type: boolean
          title: Simulate Scoring Timeout
          default: false
      type: object
      required:
      - customer_id
      - requested_amount_vnd
      - customer_age
      - is_existing_salaried
      - payment_account_id
      title: SubmitApplicationRequest
    SubmitApplicationResponse:
      properties:
        application_id:
          type: string
          title: Application Id
        state:
          type: string
          title: State
        loan_offer:
          $ref: '#/components/schemas/LoanOfferResponse'
      type: object
      required:
      - application_id
      - state
      - loan_offer
      title: SubmitApplicationResponse
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError